removed duplicate factory code
